refactor(cli): extract prompt stash logic to hook and polish UI layout
